analysisWorker/requirements.txt
psycopg2-binary>=2.9
sentence-transformers>=2.7
torch>=2.2
numpy>=1.26
